requirements.txt
# Core ML
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Technical Analysis
ta>=0.10.2

# Google Cloud Storage
google-cloud-storage>=2.10.0

# Development
pytest>=7.4.0
ruff>=0.1.0
jupyter>=1.0.0
jupyterlab>=4.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Serialization
joblib>=1.3.0
